+++ libraries/dataset/tcc/RowDataset.tcc
@@ -30,31 +30,31 @@ namespace dataset
     }
 
     template <typename ExampleType>
-    ExampleType& RowDataset<ExampleType>::GetExample(uint64_t index)
+    ExampleType& RowDataset<ExampleType>::GetExample(size_t index)
     {
         return _examples[index];
     }
 
     template <typename ExampleType>
-    const ExampleType& RowDataset<ExampleType>::GetExample(uint64_t index) const
+    const ExampleType& RowDataset<ExampleType>::GetExample(size_t index) const
     {
         return _examples[index];
     }
 
     template <typename ExampleType>
-    ExampleType& RowDataset<ExampleType>::operator[](uint64_t index)
+    ExampleType& RowDataset<ExampleType>::operator[](size_t index)
     {
         return _examples[index];
     }
 
     template <typename ExampleType>
-    const ExampleType& RowDataset<ExampleType>::operator[](uint64_t index) const
+    const ExampleType& RowDataset<ExampleType>::operator[](size_t index) const
     {
         return _examples[index];
     }
 
     template <typename ExampleType>
-    typename RowDataset<ExampleType>::Iterator RowDataset<ExampleType>::GetIterator(uint64_t fromRowIndex, uint64_t size) const
+    typename RowDataset<ExampleType>::Iterator RowDataset<ExampleType>::GetIterator(size_t fromRowIndex, size_t size) const
     {
         size = CorrectRangeSize(fromRowIndex, size);
         return utilities::MakeStlIterator(_examples.cbegin() + fromRowIndex, _examples.cbegin() + fromRowIndex + size);
@@ -63,7 +63,7 @@ namespace dataset
     template <typename ExampleType>
     void RowDataset<ExampleType>::AddExample(ExampleType example)
     {
-        uint64_t size = example.GetDataVector().Size();
+        size_t size = example.GetDataVector().Size();
         _examples.push_back(std::move(example));
 
         if (_maxExampleSize < size)
@@ -107,7 +107,7 @@ namespace dataset
     }
 
     template <typename ExampleType>
-    void RowDataset<ExampleType>::RandomSwap(std::default_random_engine& rng, uint64_t targetExampleIndex, uint64_t rangeFirstIndex, uint64_t rangeSize)
+    void RowDataset<ExampleType>::RandomSwap(std::default_random_engine& rng, size_t targetExampleIndex, size_t rangeFirstIndex, size_t rangeSize)
     {
         using std::swap;
         rangeSize = CorrectRangeSize(rangeFirstIndex, rangeSize);
@@ -123,7 +123,7 @@ namespace dataset
 
     template <typename ExampleType>
     template <typename SortKeyType>
-    void RowDataset<ExampleType>::Sort(SortKeyType sortKey, uint64_t fromRowIndex, uint64_t size)
+    void RowDataset<ExampleType>::Sort(SortKeyType sortKey, size_t fromRowIndex, size_t size)
     {
         size = CorrectRangeSize(fromRowIndex, size);
 
@@ -136,18 +136,18 @@ namespace dataset
 
     template <typename ExampleType>
     template <typename PartitionKeyType>
-    void RowDataset<ExampleType>::Partition(PartitionKeyType partitionKey, uint64_t fromRowIndex, uint64_t size)
+    void RowDataset<ExampleType>::Partition(PartitionKeyType partitionKey, size_t fromRowIndex, size_t size)
     {
         size = CorrectRangeSize(fromRowIndex, size);
         std::partition(_examples.begin() + fromRowIndex, _examples.begin() + fromRowIndex + size, partitionKey);
     }
 
     template <typename ExampleType>
-    void RowDataset<ExampleType>::Print(std::ostream& os, size_t tabs, uint64_t fromRowIndex, uint64_t size) const
+    void RowDataset<ExampleType>::Print(std::ostream& os, size_t tabs, size_t fromRowIndex, size_t size) const
     {
         size = CorrectRangeSize(fromRowIndex, size);
 
-        for (uint64_t index = fromRowIndex; index < fromRowIndex + size; ++index)
+        for (size_t index = fromRowIndex; index < fromRowIndex + size; ++index)
         {
             os << std::string(tabs * 4, ' ');
             _examples[index].Print(os);
@@ -163,7 +163,7 @@ namespace dataset
     }
 
     template <typename ExampleType>
-    uint64_t RowDataset<ExampleType>::CorrectRangeSize(uint64_t fromRowIndex, uint64_t size) const
+    size_t RowDataset<ExampleType>::CorrectRangeSize(size_t fromRowIndex, size_t size) const
     {
         if (size == 0 || fromRowIndex + size > _examples.size())
         {
